requirements.txt
Flask
gunicorn
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
